order card,edit db price , order , payment

+++ packages/db/src/migrations/20250721010601-create-products.ts
@@ -21,7 +21,7 @@ export default {
       },
       name: { type: DataTypes.STRING(191), allowNull: false },
       description: { type: DataTypes.TEXT, allowNull: true },
-      price: { type: DataTypes.DECIMAL(12, 2), allowNull: false },
+      price: { type: DataTypes.INTEGER, allowNull: false },
       stock_quantity: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
       status: { type: DataTypes.ENUM(...Object.values(ProductStatus)), allowNull: false, defaultValue: ProductStatus.ACTIVE },
       created_at: { type: DataTypes.DATE, allowNull: false, defaultValue: DataTypes.NOW },
